pyproject.toml
[tool.poetry]
name = "CV12"
version = "0.1.0"
description = ""
authors = ["CV12 <https://github.com/boostcampaitech7/level1-imageclassification-cv-12.git>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.13"
pytorch-lightning = "^2.4.0"
torch = { version = "^2.4.0", extras = ["cuda12"] }
torchvision = { version = "^0.19.0", extras = ["cuda12"] }
omegaconf = "^2.3.0"
numpy = "^2.1.0"
pandas = "^2.2.2"
pillow = "^10.4.0"
requests = "^2.32.3"
dvc = "^3.54.1"
pytest = "^8.3.2"
torchmetrics = "^1.4.1"
wandb = "^0.17.8"
scikit-learn = "^1.5.1"
pytz = "^2024.2"
jupyter-client = "^8.6.3"
jupyter-core = "^5.7.2"
matplotlib-inline = "^0.1.7"
timm = "^0.9.5"  # TIMM 추가
opencv-python = "^4.8.0"  # OpenCV 추가
torchcontrib = "^0.0.2"  # torchcontrib(SWA 포함)
efficientnet-pytorch = "^0.7.1"  # EfficientNet-PyTorch 추가
albumentations = "^1.4.16"  # Albumentations 추가

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"